chat

+++ app/Http/Controllers/PermissionController.php
@@ -13,15 +13,24 @@ class PermissionController extends Controller
      * Display a listing of the resource.
      *
      * @return \Illuminate\Http\Response
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function index(Request $request)
     {
+        $this->authorize('view_permission');
+
         $permission = Permission::query()
         ->when($request->search, function ($q, $value) use ($request){
             $q->where('name', 'like', "%".$value . "%");
         })->paginate();
         return Inertia::render('Permission/Index', [
-            'permissions' => $permission
+            'permissions' => $permission,
+            'can' => [
+                'create' => auth()->user()->can('create_permission'),
+                'update' => auth()->user()->can('update_permission'),
+                'delete' => auth()->user()->can('delete_permission'),
+                'view' => auth()->user()->can('view_permission'),
+            ]
         ]);
     }
 
@@ -29,9 +38,12 @@ class PermissionController extends Controller
      * Show the form for creating a new resource.
      *
      * @return \Illuminate\Http\Response
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function create()
     {
+        $this->authorize('create_permission');
+
         $permission = new Permission();
         $modules = $this->module();
         return Inertia::render('Permission/Create',[
@@ -43,11 +55,14 @@ class PermissionController extends Controller
     /**
      * Store a newly created resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
+     * @param \Illuminate\Http\Request $request
      * @return \Illuminate\Http\Response
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function store(Request $request)
     {
+        $this->authorize('create_permission');
+
         $request->validate([
             'name' => 'required',
             'guard_name' => 'required',
@@ -64,22 +79,25 @@ class PermissionController extends Controller
     /**
      * Display the specified resource.
      *
-     * @param  int  $id
+     * @param int $id
      * @return \Illuminate\Http\Response
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function show($id)
     {
-        //
+        $this->authorize('view_permission');
     }
 
     /**
      * Show the form for editing the specified resource.
      *
-     * @param  int  $id
+     * @param int $id
      * @return \Illuminate\Http\Response
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function edit($id)
     {
+        $this->authorize('update_permission');
         $permission = Permission::find($id);
         $modules = $this->module();
         return Inertia::render('Permission/Edit', [
@@ -91,12 +109,14 @@ class PermissionController extends Controller
     /**
      * Update the specified resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
+     * @param \Illuminate\Http\Request $request
+     * @param int $id
      * @return \Illuminate\Http\Response
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function update(Request $request, $id)
     {
+        $this->authorize('update_permission');
         $request->validate([
             'name' => 'required',
             'guard_name' => 'required',
@@ -113,11 +133,13 @@ class PermissionController extends Controller
     /**
      * Remove the specified resource from storage.
      *
-     * @param  int  $id
+     * @param int $id
      * @return \Illuminate\Http\Response
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function destroy($id)
     {
+        $this->authorize('delete_permission');
         $role = Permission::find($id)->delete();
         return redirect()->route('permission.index')->withFlash('success', "Permission successfully removed");
     }
